Move PropertyListStorageURL to the app target
Issue #238
