+++ TelephoneTests/PropertyListStorageURLTests.swift
@@ -1,5 +1,5 @@
 //
-//  PropertyListStorageURL.swift
+//  PropertyListStorageURLTests.swift
 //  Telephone
 //
 //  Copyright © 2008-2016 Alexey Kuznetsov
@@ -17,11 +17,14 @@
 //
 
 import Foundation
+import XCTest
 
-public struct PropertyListStorageURL {
-    public let url: URL
+final class PropertyListStorageURLTests: XCTestCase {
+    func testUsesNameAsFileNameAndPlistAsExtension() {
+        let name = "any-name"
 
-    public init(directory: URL, name: String) {
-        url = directory.appendingPathComponent("\(name).plist")
+        let sut = PropertyListStorageURL(directory: URL(fileURLWithPath: NSTemporaryDirectory()), name: name)
+
+        XCTAssertEqual(sut.url.lastPathComponent, "\(name).plist")
     }
 }
